Move EEPROM reset check to main loop
- Instead of checking for an EEPROM reset request at startup, check each
pass through the loop. EPROM reset is signaled by placing and removing a
jumper, or by momentarily closing a switch. This simplifies the reset
process once the multigame is fully installed. A switch can be brought
out near the thruster behind the cabinet, as part of the thruster
assembly, etc.

- Add function headers.